\name{deque<<%= ctype_r %>>-class}
\Rdversion{1.1}
\docType{class}
\alias{deque<<%= ctype_r %>>-class}

\title{Class "deque<<%= ctype_r %>>"}
\description{wrapper around the \code{std::deque<<%= ctype_r %>>} c++ class}
\section{Objects from the Class}{
Objects can be created by calls of the form 
\code{new( CPP("deque<<%= ctype_r %>>" ) )}.
}
\section{Slots}{
  \describe{
  	  \item{\code{pointer}:}{external pointer to the C++ \code{std::deque<<%= ctype_r %>>} object }
  }
}
\section{Extends}{
Class \code{"\linkS4class{deque<>}"}, directly.
Class \code{"\linkS4class{C++Object}"}, by class "deque<>", distance 2.
}
\section{Methods}{
No methods defined with class "deque<<%= ctype_r %>>" in the signature.
}
\references{ The \code{std::deque} C++ template from the standard template library.  }
\author{ Romain Francois <francoisromain@free.fr> }
\seealso{ \linkS4class{deque<>} for examples }
\keyword{classes}
