\name{set<<%= ctype_r %>>-class}
\Rdversion{1.1}
\docType{class}
\alias{set<<%= ctype_r %>>-class}

\title{Class "set<<%= ctype_r %>>" ~~~ }
\description{wrapper around the \code{std::set<<%= ctype_r %>>} c++ class}
\section{Objects from the Class}{
Objects can be created by calls of the form 
\code{new( CPP("set<<%= ctype_r %>>" ) )}.
}
\section{Slots}{
  \describe{
  	  \item{\code{pointer}:}{external pointer to the C++ \code{std::set<<%= ctype_r %>>} object }
  }
}
\section{Extends}{
Class \code{"\linkS4class{set<>}"}, directly.
Class \code{"\linkS4class{C++Object}"}, by class "set<>", distance 2.
}
\section{Methods}{
No methods defined with class "set<<%= ctype_r %>>" in the signature.
}
\references{ The \code{std::set} C++ template from the standard template library.  }
\author{ Romain Francois <francoisromain@free.fr> }
\seealso{ \linkS4class{set<>} for examples }
\keyword{classes}
